Fix comparisons (with cast)

+++ src/org/cryptonit/CryptonitApplet.java
@@ -495,8 +495,8 @@ public class CryptonitApplet extends Applet implements ExtendedLength {
         byte p1 = buf[ISO7816.OFFSET_P1];
         byte p2 = buf[ISO7816.OFFSET_P2];
 
-        if ((p1 == 0x07 || p1 == 0x11 || p1 == 0x14)
-                && (p2 == 0x9a || p2 == 0x9c || p2 == 0x9d)) {
+        if ((p1 == (byte) 0x07 || p1 == (byte) 0x11 || p1 == (byte) 0x14)
+                && (p2 == (byte) 0x9a || p2 == (byte) 0x9c || p2 == (byte) 0x9d)) {
             doPrivateKeyOperation(apdu);
         } else {
             ISOException.throwIt(ISO7816.SW_INCORRECT_P1P2);
